Moved and renamed _SearchAgain.js to BookCard.js. Updated components to use renamed BookCard.js
